+++ test/cmd/kuberc.sh
@@ -26,18 +26,22 @@ run_kuberc_tests() {
   kube::log::status "Testing kubectl alpha kuberc set commands"
 
   KUBERC_FILE="${TMPDIR:-/tmp}"/kuberc_file
+  cat > "$KUBERC_FILE" << EOF
+apiVersion: kubectl.config.k8s.io/v1beta1
+kind: Preference
+EOF
 
   # Build up the kuberc file using kubectl alpha kuberc set commands
-  kubectl alpha kuberc set --file="$KUBERC_FILE" --section=defaults --command=apply --option=server-side=true --option=dry-run=server --option=validate=strict
-  kubectl alpha kuberc set --file="$KUBERC_FILE" --section=defaults --command=delete --option=interactive=true
-  kubectl alpha kuberc set --file="$KUBERC_FILE" --section=defaults --command=get --option=namespace=test-kuberc-ns --option=output=json
+  kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=defaults --command=apply --option=server-side=true --option=dry-run=server --option=validate=strict
+  kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=defaults --command=delete --option=interactive=true
+  kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=defaults --command=get --option=namespace=test-kuberc-ns --option=output=json
 
-  kubectl alpha kuberc set --file="$KUBERC_FILE" --section=aliases --name=crns --command="create namespace" --appendarg=test-kuberc-ns
-  kubectl alpha kuberc set --file="$KUBERC_FILE" --section=aliases --name=getn --command=get --prependarg=namespace --option=output=wide
-  kubectl alpha kuberc set --file="$KUBERC_FILE" --section=aliases --name=crole --command="create role" --option=verb=get,watch
-  kubectl alpha kuberc set --file="$KUBERC_FILE" --section=aliases --name=getrole --command=get --option=output=json
-  kubectl alpha kuberc set --file="$KUBERC_FILE" --section=aliases --name=runx --command=run --option=image=nginx --option=labels=app=test,env=test --option=env=DNS_DOMAIN=test --option=namespace=test-kuberc-ns --appendarg=test-pod-2 --appendarg=-- --appendarg=custom-arg1 --appendarg=custom-arg2
-  kubectl alpha kuberc set --file="$KUBERC_FILE" --section=aliases --name=setx --command="set image" --appendarg=pod/test-pod-2 --appendarg=test-pod-2=busybox
+  kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=aliases --name=crns --command="create namespace" --appendarg=test-kuberc-ns
+  kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=aliases --name=getn --command=get --prependarg=namespace --option=output=wide
+  kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=aliases --name=crole --command="create role" --option=verb=get,watch
+  kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=aliases --name=getrole --command=get --option=output=json
+  kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=aliases --name=runx --command=run --option=image=nginx --option=labels=app=test,env=test --option=env=DNS_DOMAIN=test --option=namespace=test-kuberc-ns --appendarg=test-pod-2 --appendarg=-- --appendarg=custom-arg1 --appendarg=custom-arg2
+  kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=aliases --name=setx --command="set image" --appendarg=pod/test-pod-2 --appendarg=test-pod-2=busybox
 
   kube::log::status "Testing kubectl alpha kuberc view commands"
   # Test: kubectl alpha kuberc view
@@ -55,25 +59,25 @@ run_kuberc_tests() {
   kube::test::if_has_string "${output_message}" "\"kind\": \"Preference\""
 
   # Test: Attempt to set existing default without --overwrite flag should fail
-  output_message=$(! kubectl alpha kuberc set --file="$KUBERC_FILE" --section=defaults --command=get --option=output=yaml 2>&1)
+  output_message=$(! kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=defaults --command=get --option=output=yaml 2>&1)
   kube::test::if_has_string "${output_message}" "defaults for command \"get\" already exist, use --overwrite to replace"
 
   # Test: Now set with --overwrite flag should succeed and merge options
-  kubectl alpha kuberc set --file="$KUBERC_FILE" --section=defaults --command=get --option=output=yaml --overwrite
+  kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=defaults --command=get --option=output=yaml --overwrite
   output_message=$(kubectl alpha kuberc view --kuberc="$KUBERC_FILE")
   kube::test::if_has_string "${output_message}" "default: yaml"
   # Should still have namespace option from before
   kube::test::if_has_string "${output_message}" "default: test-kuberc-ns"
 
   # Test: Attempt to set existing alias without --overwrite flag should fail
-  output_message=$(! kubectl alpha kuberc set --file="$KUBERC_FILE" --section=aliases --name=getn --command=get --prependarg=pods 2>&1)
+  output_message=$(! kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=aliases --name=getn --command=get --prependarg=pods 2>&1)
   kube::test::if_has_string "${output_message}" "alias \"getn\" already exists, use --overwrite to replace"
 
   # Test: Error cases - Missing required flags
-  output_message=$(! kubectl alpha kuberc set --file="$KUBERC_FILE" --command=get --option=output=wide 2>&1)
+  output_message=$(! kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --command=get --option=output=wide 2>&1)
   kube::test::if_has_string "${output_message}" "required flag(s) \"section\" not set"
 
-  output_message=$(! kubectl alpha kuberc set --file="$KUBERC_FILE" --section=defaults --option=output=wide 2>&1)
+  output_message=$(! kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=defaults --option=output=wide 2>&1)
   kube::test::if_has_string "${output_message}" "required flag(s) \"command\" not set"
 
   # Test: KUBERC=off with view command
@@ -85,9 +89,9 @@ run_kuberc_tests() {
   kube::test::if_has_string "${output_message}" "KUBERC is disabled via KUBERC=off environment variable"
 
   # Restore getn alias back to "namespace" for remaining tests
-  kubectl alpha kuberc set --file="$KUBERC_FILE" --section=aliases --name=getn --command=get --prependarg=namespace --option=output=wide --overwrite
+  kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=aliases --name=getn --command=get --prependarg=namespace --option=output=wide --overwrite
   # Restore get defaults back to namespace=test-kuberc-ns and output=json for remaining tests
-  kubectl alpha kuberc set --file="$KUBERC_FILE" --section=defaults --command=get --option=namespace=test-kuberc-ns --option=output=json --overwrite
+  kubectl alpha kuberc set --kuberc="$KUBERC_FILE" --section=defaults --command=get --option=namespace=test-kuberc-ns --option=output=json --overwrite
 
   kube::log::status "Testing kuberc aliases and defaults functionality"
